Merged PR 82126: Added script to generate test vector for HMAC dynamically
When UVM test is run, the driver calls the python script to run OpenSSL commands to generate random key, message and corresponding digest. These are stored in a text file that the driver then parses to queue up write commands to the DUT. Message can be of any length between 1 and 512 bytes. The predictor also reads the file to capture the expected result. Caveat is that the text file is overwritten for every test so we only have the last vector if we want to debug later.

Related work items: #295325