Merged PR 83787: Add Interrupt RegFile template
# Overview
Template for implementing interrupt control registers in each component.
This template implements 4 notification interrupts (notif0-notif3) and 4 error interrupts (error0-error3). The names and numbers of these events is expected to vary for each peripheral and must be manually changed and connected as applicable.
The RDL reg_gen script is updated to generate bit masks and indices for each register field, and adds an addr_width param to the generated package for more robust instantiation.
This PR also includes Interrupts implemented in the Mailbox; you may review mbox_reg.rdl for an example of how the template may be inserted into an extant register block.
As a result of adding support for mailbox interrupts, this PR is able to remove the sim_irq_gen model that is non-synthesizable (along with the TB routines to manage it).

# Alignment of registers is:
0x0 - Global or statically implemented registers
0x100 - Scalable array of interrupt statistics counters, one 32-bit counter for each event to report number of occurrences.
0x200 - Implemented as a set of triggers for the statistics counters; due to RDL restrictions this had to be added to the address map but should be treated by engineers as a reserved region in memory and not accessed by SW (it is also R-only).

Related work items: #297645, #297889, #301294, #301296, #301298, #301299