Merged PR 85530: HMAC lint clean up
Fixed a number of lint errors and warnings in HMAC, KV, ahb_slv_sif and sha512_core.

There are a few more warnings primarily in KV. There are 8 errors in hmac related to signals assigned but not used - these are unused signals and seems ok to waive. I have created a waiver file that is included here.

Note about waivers: There is a bug in the current pb fe lint rtl flow that causes the waivers to not to applied. Working to CAD team to get that resolved.

Related work items: #396911