[entropy_src] Latch rng_enable
Upstream OpenTitan commit 96b75ae6dadd76056ead052922e034cce016e53b

rng_enable is being used in other clock domains. The logic did not latch
the signal in the previous design.

This commit latches the signal to make it a glitch-free signal.

In the entropy_src top, the RTL seems to have a mux. However, `Stub` is
a parameter so the mux will be optimized out.

Signed-off-by: Eli Kim <eli@opentitan.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>