Merged PR 92906: Fixing mailbox bug where dataout register doesn't update fast enough
Fixing the dataout register functionality.

Now it will update with the first entry on the arc to execute.
Each subsequent read will trigger an update to the register.
If the update is in progress when another read is being made, we hold the request until the update completes.

Related work items: #425793