Adding csrng/entropy_src to caliptra_top (#39)
* Adding trng to caliptra_top

* Fixing range issue in physical_rng and adding vcs to Makefile

* Update top level portname from trng_req -> etrng_req in hdl_top.sv

* Making itrng_data/valid fixed ports at the top level

* Fixed assertions that were firing off when reset was asserted.

The assertion was firing off a thread since rstn was 1 prior to it being
asserted to 0.  This caused the simulator to expect a valid property
three cycles later while reset was still 0.  We don't care about the
property during this time, so we add an '|| !rstn' to keep the assertion
from firing off erroneously.