[aes, rtl] Update comment on the suitability of the masking PRNG
Upstream OpenTitan commit f111d0231727a1f933433d0aca6d207e32d3d20a

This comment was added when adding the first, initial PRNG
implementation and never changed since then. In the meantime, we have
however performed many measurements to optimize the design of the PRNG
and the masking implementation. We are confident that the current
combination of PRNG and masking implementation provide sufficient SCA
resistance.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>