Merged PR 102362: New random reg access sequence for uvmf caliptra top
This sequence will randomly select addresses for the uC to read, and puts them in the mailbox.

Out of bounds addresses are sent with OOB_ACCESS command and trigger an expected NMI. If they don't FW will signal an error.

In bounds accesses are read and the data is written back to the mailbox.

Related work items: #428994