Merged PR 103317: bugfix - 455868 - dlen restricts size of data response that may be read
Receiver (uC) needs to program DLEN appropriately when populating the response.
Requester can read DLEN when processing a response in order to know the size of the response.

Snapshot of DLEN is taken at state transition that transfers control between sender/receiver so that the client reading DATAOUT cannot overwrite the DLEN that the sender wrote in order to read beyond the valid data.

Related work items: #455868