Merged PR 105207: added privkey in and out to handle fw access to secret values
PRIVKEY_IN is a register as an input for signing operation by KV or FW or PCR. FW has only a write access to this register.

PRIVKEY_OUT is a register as an output for keygen operation. It can be read by FW if the KV is not the destination for keygen operation.

ecc_uvmf and ecc_top_tb have been updated too.

Related work items: #462145