[csrng, rtl] Remove output mux on genbits bus
Upstream OpenTitan commit 646bd2db61198c63a8dc57da0f670a3392163c7a

This output mux is not functionally required but poses an FI target.
Therefore, it has been decided to remove it. Related to that, the output
assertions are modified to use the valid bit as enable term.

This related to lowRISC/OpenTitan#16564.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>