[entropy_src] Doc and check required bypass health test window size
Upstream OpenTitan commit c819ca806414dbf4a654b51173bfa925a377ff57

In boot-time mode, the entropy source generates exactly one 384-bit seed
without hardware conditioning. This requires the bypass/boot-time health
test window size to be configured to 384 bits. Other window sizes are
not tested and also not supported at the moment. The same restriction
applies to the non-boot-time bypass mode, i.e., if ES_ROUTE and ES_TYPE
are both set to True.

This commit documents this requirement in the programmers guide and the
RTL, and it adds a SystemVerilog assertion to prevent mis-configuration
during simulation.

The following files are changed in OpenTitan commit but not included:
- hw/ip/entropy_src/doc/programmers_guide.md

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>