[csrng/rtl] Restrict `glen` to 12 bit (thus <= 4095)
Upstream OpenTitan commit 477ae95158374f1c4ae940bcd66c1785d55aee1d

Prior to this commit, the `glen` field, which defines the number of
128-bit entropy block each Generate command returns, had a width of 13
bit (`[24:12]`) and thus took a maximum value of 2^13 - 1 = 8191.  This
is problematic because it would have led CSRNG to return up to 8191 *
128 bit, which is larger than the maximum number of bit per request
allowed by NIST (2^19) bit, for one Generate command (#18334).

This commit reduces the width of the `glen` field to 12 bit, so that its
maximum value is now 4095, thereby reducing the upper bound of bit per
request below the maximum stipulated by NIST.  This commit thus resolves
the HW part of #18334.

The upper bound is now 128 bit lower than the allowed maximum, which
could have been avoided if the `glen` field was kept at a width of 13
bit and CSRNG instead checked its value and returned an error if it
exceeded 2^12.  Such an error response mechanism is currently not
implemented, however, and implementing one from scratch was out of scope
for this change.

This change was tested by setting `glen` to 4095 in `csrng_smoke_test`;
the test passed and a manual inspection of the waves looked correct.
This change was further tested with a full block-level regression run,
although randomization of `glen` is currently limited to the range
[1:32]; this will be resolved separately (due to deadline constraints).

The following files are changed in OpenTitan commit but not included:
- hw/dv/sv/csrng_agent/csrng_item.sv
- hw/dv/sv/csrng_agent/csrng_monitor.sv
- hw/ip/csrng/doc/theory_of_operation.md
- hw/ip/csrng/dv/cov/csrng_cov_if.sv
- hw/ip/csrng/dv/env/csrng_scoreboard.sv
- hw/ip/csrng/dv/env/seq_lib/csrng_alert_vseq.sv

Signed-off-by: Andreas Kurth <adk@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>