Merged PR 113052: New UVM mailbox test cases - inject random delays
Add test sequences that inject random delays in mailbox flows.

Also add a fix for the mbox_dataout prediction logic that initializes the register to 0 on every mailbox flow (as is done in hw) to account for the case where no data is written to the mailbox during the command initialize phase (e.g. DLEN = 0).

Related work items: #442947, #490264