Merged PR 118100: UVM protocol error testcases
Add testcases that randomly inject invalid/illegal mailbox register accesses during a regular mailbox flow.
Testcases are added to inject invalid register accesses for both SOC->UC and UC->SOC mailbox flows in this PR. (only APB invalid reg accesses are performed, to validate the protocol error path).
Invalid register accesses are defined (as in the integration spec) as writes to an register (or reads from dataout) at an unexpected stage of the flow (i.e., write to clear execute when the responder has not yet returned control to the sender).
The soc_ifc UVM bench also includes new code to initialize interrupts and service error events when they occur to restore the mailbox state.
Also added a fix for a minor issue in the mailbox - certain fsm 'arc' signals toggled in the incorrect state (not a functional issue, but a power optimization).

Related work items: #442339, #442340, #442948, #442949, #495747, #521329