Merged PR 144445: Hand-copy file sync from GH repo - purge obsolete READMEs
Hand-copy file sync from GH repo - purge obsolete READMEs
This step is a precursor to the repository restructure that will delete many contents from the internal Microsoft repository, while adding the GH repository as a comodule that contains those identical contents.
Any post-1.0 changes will be merged simultaneously to the GitHub repository, also as a precursor to the repo restructure.
Future engineers may be interested in comparing the snapshot of GitHub against the snapshot of Microsoft's internal repository immediately before the restructuring occurred, and this PR makes that easier.

Related work items: #617002