Revert PR #438 with shortened SHA test vector
This fix is unnecessary because the failure scenario is already
accounted for by a previous fix that uses a C macro/preprocessor
conditional to select an appropriate test vector for the sha accel smoke
test based on the Caliptra configuration (Core vs Seed).
Leaving a large vector in place is advantageous for testing the Core
mode to ensure large messages work as expected, though the short vector
is required for the Seed smoke test in order to compile the code.