[UVM] Fix error detection in certain error injection tests (#467)
* Enhance error prediction abilities for error injection mailbox sequence

Add different types of errors that may be expected when injecting errors
to a sequence.
In the handler sequence, detect when the requested response dlen is out
of bounds and flag a cmd failure.
In the SoC Mailbox command issuer sequence, correctly set the expected
error type. When random reg writes overwrite the requested response
dlen, flag that type of error. In the status evaluator, allow
CMD_FAILURE even without an error interrupt for this event, because
there is no protocol violation to expect.

* Flag a CMD_FAILURE when requested resp_dlen is out of bounds

* Demote error to 'info' regarding nop writes to CPTRA_FW_EXTENDED_ERROR_INFO