[TESTS] SHA3 external mu test (#956)
* [TESTS] External mu test uses SHA3 block

This creates a check_external_mu function inside the
smoke_test_mldsa_external_mu test. Before invoking the external mu flow
in Adam's bridge it now calculates the external mu with the SHA3 engine
and compares this to the pre-computed value. If these match then it goes
through the normal MLDSA external mu flow as before.

* [TESTS] MLDSA smoke test fix vectors

Before these test vectors were reversed before going through the MLDSA
flow which adds about 10 minutes of simulation time on Verilator. This
commit reverses the constant vectors to avoid wasting this time. The
commit adjusts both the MLDSA test as well as the external mu variant.

* [TESTS] Message endianness now configurable

Also change the calls to the SHA3 start function in cshake, sha3,
interrupt and external mu tests.

* [TESTS] SHA3 external mu test added

This runs ACVP NIST KAT vectors through the newly added SHA3 block. The
test runs 12 vectors through the SHA3 block with public key and
processed message and compares it to a pre-computed value made by
dilithium-py.

Here are the original vectors:
https://raw.githubusercontent.com/usnistgov/ACVP-Server/refs/heads/master/gen-val/json-files/ML-DSA-sigGen-FIPS204/internalProjection.json

* [TESTS] Confusing comment removed

* [TB] External mu forcing services

This implements 4 separate forcing sequences. These are necessary for
full random external mu testing. Here are the codes:
- 0xD0: force the public key into the SHA3 message FIFO.
- 0xD1: force the message into the SHA3 message FIFO.
- 0xD2: start keygen and signing for ABR without injecting message.
- 0xD3: start verification without injecting message.

Co-authored-by: Gary Guo <gary.guo@lowrisc.org>

* [TESTS] External mu full random test

This exercises the new commands added in the testbench services and is
based on the smoke_test_mldsa_keygen_sign_vfy_rand test.

Move the initialisation of the interrupts to after we've done forcing
signals inside the SHA3 block.

* [TB] Trailing whitespace removed from services

* [TESTS] External mu added to random nightly

* [TESTS] SHA3 interrupts added to directed tests

* [TESTS] SHA3 Caliptra ISR updates

For the tests using SHA3 add functionality to deal with interrupts.

* Revert "[TESTS] SHA3 and CSHAKE smoke tests no notif"

This reverts commit 98fe31cc8d9d27b9b91d3dcb4c9b7375e6ab9a57.

* [TESTS] External mu rand test prints

Add prints to help find the progression of the external mu random test.

* [TB] Parameterise MLDSA test vector

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_externalmu' with updated timestamp and hash after successful run

---------

Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Co-authored-by: Antonio Martinez Zambrana (LOWRISC C I C) <v-amartinezz@microsoft.com>